requirements_dev.txt
plotly==5.11.0
ipywidgets==7.6
jupyterlab
matplotlib

numpy
gmsh
pythreejs==2.4.2
git+https://github.com/OpenOrion/su2fmt.git#egg=su2fmt
shapely
scipy
ipython_genutils
jupyter_cadquery